Move override files to '/schedules/calendar' directory
